requirements.txt
kafka-python==2.0.2
pandas==2.2.3
scikit-learn==1.4.2
streamlit==1.35.0
